Reorganize notifications.py tests
* This is a nice compromise between using unittest.TestCase and Pytest fixtures that
allows for better organization, shorter names, less places to rename in the event of src
rename, and provides much better context
